+++ NEWS.md
@@ -1,5 +1,8 @@
 # campfin (development version)
 
+* The `end` argument of `str_normal()` can now be controlled with `abb_end` in
+  `normal_address()` (default `TRUE`).
+
 # campfin 1.0.8
 
 * Add `string` argument to `guess_delim()` to read the input as a single line